Add short url creation capabilities using bit.ly
This will comb through the html files of your blog and replace the following string SHORTURLHERE with a bitl.ly link for the page itself. It also does the same in the Markdown file for the article.

I created this so that I could have a way for people to tweet snippets from my blog, with the included short-url in the tweet. 

This is how I will use it for example: 

TEXT-YOU-WANT-TWEETED <a href="http://twitter.com/home/?status= TEXT-YOU-WANT-TWEETED. via SHORTURLHERE @justinchaschowy " target="_blank"> [tweet this].</a>